Documentation changes, modified set-secret.sh to use ruby instead of rake

+++ src/set-secret.sh
@@ -1,7 +1,12 @@
 #!/bin/bash
 
-rm ~/Dev/spaceman/config/secrets.yml
+if [ -e ~/Dev/spaceman/config/secrets.yml ]; then
+	rm ~/Dev/spaceman/config/secrets.yml
+fi
+
+key=`ruby -e "require 'securerandom'; puts SecureRandom.hex(64)"`
+
 echo "production:" > ~/Dev/spaceman/config/secrets.yml
-echo "  secret_key_base: `bundle exec rake secret`" >> ~/Dev/spaceman/config/secrets.yml
+echo "  secret_key_base: ${key}" >> ~/Dev/spaceman/config/secrets.yml
 
 echo "Done"
